change double to in Account

+++ commonData/src/ch/hsr/objectCaching/account/AccountImpl.java
@@ -4,13 +4,13 @@ package ch.hsr.objectCaching.account;
 
 public class AccountImpl implements Account {
 	
-	private int balance;
+	private double balance;
 
-	public int getBalance() {
+	public double getBalance() {
 		return balance;
 	}
 	
-	public void setBalance(int balance)
+	public void setBalance(double balance)
 	{
 		this.balance = balance;
 	}
